roomclip-scraping/pyproject.toml
[tool.poetry]
name = "roomclip-scraping"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.3"
pymysql = "^1.0.2"
cryptography = "^39.0.2"
flask-httpauth = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
